Extract AttributeParse to make rubocop happy
